requirements.txt
asgiref==3.6.0
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
mysqlclient==2.1.1
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
